ui: factor recps out of composer and remove drafts
